+++ scim2-sdk-client/src/main/java/com/unboundid/scim2/client/requests/ResourceReturningRequestBuilder.java
@@ -37,7 +37,7 @@ public abstract class ResourceReturningRequestBuilder
   protected boolean excluded;
 
   /**
-   * The attribute list of include or exclude.
+   * The attribute list to include or exclude.
    */
   @Nullable
   protected Set<String> attributes;