+++ CHANGELOG.md
@@ -12,6 +12,9 @@ set the following property in your application:
 PatchOperation.APPEND_NEW_PATCH_VALUES_PROPERTY = true;
 ```
 
+Updated `SearchRequestBuilder` to be more permissive of ListResponses with non-standard attribute
+casing (e.g., if a response includes a `"resources"` array instead of `"Resources"`).
+
 ## v3.2.0 - 2024-Dec-04
 Fixed an issue where `AndFilter.equals()` and `OrFilter.equals()` could incorrectly evaluate to
 true.